[Chore] 말풍선 컴포넌트 분리
