+++ src/components/live-view/LiveIntro.tsx
@@ -1,8 +1,5 @@
-import { ReactNode } from 'react';
-import { IoTriangleSharp } from 'react-icons/io5';
-
-import { twMerge } from '@/lib/twMerge';
 import Heading2 from '@components/common/program/program-detail/Heading2';
+import SpeechBubble from '@components/common/program/program-detail/SpeechBubble';
 import SuperTitle from '@components/common/program/program-detail/SuperTitle';
 
 const superTitle = [
@@ -96,44 +93,6 @@ function LiveIntro() {
         </div>
       </div>
     </section>
-  );
-}
-
-function SpeechBubble({
-  className,
-  children,
-  bgColor,
-  tailPosition = 'right',
-  tailHidden = false,
-}: {
-  className?: string;
-  children?: ReactNode;
-  bgColor?: string;
-  tailPosition?: 'left' | 'right';
-  tailHidden?: boolean;
-}) {
-  return (
-    <div
-      className={twMerge(
-        'min-w-50 relative w-fit rounded-full bg-neutral-90 px-8 py-4 text-xxsmall12 font-semibold text-neutral-0',
-        className,
-      )}
-      style={{ backgroundColor: bgColor }}
-    >
-      {children}
-      {!tailHidden && (
-        <IoTriangleSharp
-          className={twMerge(
-            'absolute -bottom-4',
-            tailPosition === 'right'
-              ? 'right-8 -rotate-90'
-              : 'left-8 rotate-90',
-          )}
-          size={32}
-          color={bgColor ? bgColor : '#F3F3F3'}
-        />
-      )}
-    </div>
   );
 }
